test: add comprehensive test suites (119 cases), fix frame parser sign extension
- test_crypto: 20 cases (Base64 full vectors, SHA1 vectors, WebSocket accept key)
- test_frame: 22 cases (all opcodes, mask/unmask, extended lengths, edge cases)
- test_ringbuffer: 14 cases (push/peek/advance, wrap, iovec, stress cycle)
- test_pool: 15 cases (ObjectPool acquire/release, ServerStats overload)
- test_vocabulary: 41 cases (expected, optional, FixedVector, FixedString, FixedFunction, ScopeGuard)
- test_utils: 7 cases (legacy, kept for compatibility)
- Fix char sign extension in parse_frame_header (uint8_t cast)
- Remove ASIO references from README
- Update test badge to 119 passed

+++ include/ewss/utils.hpp
@@ -250,17 +250,19 @@ inline size_t parse_frame_header(std::string_view data, FrameHeader& header) {
 
   if (len == 126) {
     if (data.size() < 4) return 0;
-    len = (static_cast<uint16_t>(data[2]) << 8) | data[3];
+    len = (static_cast<uint64_t>(static_cast<uint8_t>(data[2])) << 8) |
+          static_cast<uint64_t>(static_cast<uint8_t>(data[3]));
     header_size = 4;
   } else if (len == 127) {
     if (data.size() < 10) return 0;
-    len = (static_cast<uint64_t>(data[2]) << 56) |
-          (static_cast<uint64_t>(data[3]) << 48) |
-          (static_cast<uint64_t>(data[4]) << 40) |
-          (static_cast<uint64_t>(data[5]) << 32) |
-          (static_cast<uint64_t>(data[6]) << 24) |
-          (static_cast<uint64_t>(data[7]) << 16) |
-          (static_cast<uint64_t>(data[8]) << 8) | data[9];
+    len = (static_cast<uint64_t>(static_cast<uint8_t>(data[2])) << 56) |
+          (static_cast<uint64_t>(static_cast<uint8_t>(data[3])) << 48) |
+          (static_cast<uint64_t>(static_cast<uint8_t>(data[4])) << 40) |
+          (static_cast<uint64_t>(static_cast<uint8_t>(data[5])) << 32) |
+          (static_cast<uint64_t>(static_cast<uint8_t>(data[6])) << 24) |
+          (static_cast<uint64_t>(static_cast<uint8_t>(data[7])) << 16) |
+          (static_cast<uint64_t>(static_cast<uint8_t>(data[8])) << 8) |
+          static_cast<uint64_t>(static_cast<uint8_t>(data[9]));
     header_size = 10;
   }
